+++ src/Core/CleanArc.Domain/Entities/User/UserClaim.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace CleanArc.Domain.Entities.User;
 
-public class UserClaim:IdentityUserClaim<int>,IEntity
+public class UserClaim:IdentityUserClaim<long>,IEntity
 {
     public User User { get; set; }
 }
